refactor: 엔티티 정적 팩토리 메서드 추가 및 nullable=false 필드 primitive 타입 변경 - Team, TeamMembers, Challenge, ChallengeWeek, ChallengeCategory에 정적 팩토리 메서드 추가 - nullable=false 컬럼 Integer/Long → int/long 변경 - Team에서 minMemberCount 제거 (고정 정책이므로 DTO 검증으로 대체) - ChallengeRepository에 findByTeamAndStatus 쿼리 추가 - ErrorCode에 CHALLENGE_NOT_JOINABLE, DUPLICATE_CATEGORY 추가

+++ src/main/java/com/jjanpot/server/domain/challenge/entity/ChallengeMinGoalPolicy.java
@@ -1,5 +1,7 @@
 package com.jjanpot.server.domain.challenge.entity;
 
+import org.hibernate.annotations.Comment;
+
 import com.jjanpot.server.global.entity.BaseEntity;
 
 import jakarta.persistence.Column;
@@ -13,7 +15,6 @@ import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
-import org.hibernate.annotations.Comment;
 
 @Entity
 @Table(name = "challenge_min_goal_policy")
@@ -31,9 +32,9 @@ public class ChallengeMinGoalPolicy extends BaseEntity {
 
 	@Column(name = "member_count", nullable = false, unique = true)
 	@Comment("인원 수")
-	private Integer memberCount;
+	private int memberCount;
 
 	@Column(name = "min_amount", nullable = false)
 	@Comment("최소 목표 금액")
-	private Long minAmount;
+	private int minAmount;
 }
